+++ examples/experimental/dagster-airlift/dagster_airlift/core/sensor/__init__.py
@@ -1,0 +1,5 @@
+from .builder import (
+    DEFAULT_AIRFLOW_SENSOR_INTERVAL_SECONDS as DEFAULT_AIRFLOW_SENSOR_INTERVAL_SECONDS,
+    AirflowPollingSensorCursor as AirflowPollingSensorCursor,
+    build_airflow_polling_sensor_defs as build_airflow_polling_sensor_defs,
+)
